refactor: modifying the design and running pre-commit linters
