+++ pyproject.toml
@@ -65,7 +65,7 @@ addopts = [
   "--benchmark-warmup-iterations=100000",
   "--benchmark-disable-gc",
   "--benchmark-histogram=bench-hists/bench",
-  '--cov=qiboml',
+  '--cov=mpstab',
   '--cov-append',
   '--cov-report=xml',
   '--cov-report=html',
